#44 Move cam controllers into the new modeController directory, move speedTracker into local.
