--- app/cameraControllers/index.js
@@ -1,9 +1,0 @@
-import shipPilot from "./shipPilot";
-import freeCam from "./freeCam";
-import godCam from "./godCam";
-
-export default [
-  shipPilot,
-  freeCam,
-  godCam,
-];
